requirements.txt
motor
python-dotenv
fastapi
pydantic